+++ bin/faucet/src/server/resources/index.js
@@ -183,8 +183,12 @@ document.addEventListener('DOMContentLoaded', function () {
             } else {
                 importCommand.style.display = 'none';
             }
-            txLink.href = data.explorer_url + '/tx/' + data.transaction_id;
+
             txLink.textContent = data.transaction_id;
+            // If the explorer URL is available, set the link.
+            if (data.explorer_url) {
+                txLink.href = data.explorer_url + '/tx/' + data.transaction_id;
+            }
         });
     }