+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added support for TLS to faucet's connection to node RPC (#976).
 - Replaced integer-based duration args with human-readable duration strings (#998).
 - [BREAKING] Refactor the `miden-proving-service` proxy status service to use gRPC instead of HTTP (#953).
+- Added configurable network id for the faucet (#1016).
 
 ### Changes